Extrat TestRequestBase.
